+++ bfg9000/tools/jvm.py
@@ -337,7 +337,7 @@ class JvmPackageResolver:
 
     # TODO: Remove headers/libs from arguments after 0.7 is released.
     def resolve(self, name, submodules, version, kind, *, headers=None,
-                libs=None):
+                libs=None, system=True):
         lib = self._library(name)
         pkg = JvmPackage(name, self.builder.object_format, libs=[lib])
         log.info('found package {!r} via path-search in {!r}'