+++ test/unit/tools/msvc/test_msvc.py
@@ -4,6 +4,7 @@ from ... import *
 from .common import known_langs, mock_execute, mock_which
 
 from bfg9000 import options as opts
+from bfg9000.file_types import Directory, HeaderDirectory
 from bfg9000.options import option_list
 from bfg9000.packages import PackageKind
 from bfg9000.path import Path
@@ -16,10 +17,14 @@ def mock_execute_pkgconf(args, **kwargs):
         return '1.2.3\n'
     elif '--variable=pcfiledir' in args:
         return '/path/to/pkg-config\n'
-    elif '--cflags' in args or '--cflags-only-I' in args:
+    elif '--cflags-only-I' in args:
         return '/I/path\n' if '--msvc-syntax' in args else '-I/path\n'
+    elif '--cflags-only-other' in args:
+        return '/DMACRO\n' if '--msvc-syntax' in args else '-DMACRO\n'
     elif '--libs-only-L' in args:
         return '/LIBPATH:/path\n' if '--msvc-syntax' in args else '-L/path\n'
+    elif '--libs-only-other' in args:
+        return '/DEBUG\n' if '--msvc-syntax' in args else '-DEBUG\n'
     elif '--libs-only-l' in args:
         return 'foo.lib\n' if '--msvc-syntax' in args else '-lfoo\n'
     elif '--variable=install_names' in args:
@@ -161,10 +166,12 @@ class TestMsvcPackageResolver(CrossPlatformTestCase):
 
     def check_package(self, pkg):
         self.assertEqual(pkg.name, 'foo')
-        self.assertEqual(pkg.compile_options(self.compiler),
-                         option_list('/I/path'))
+        self.assertEqual(pkg.compile_options(self.compiler), option_list(
+            '/DMACRO', opts.include_dir(HeaderDirectory(Path('/path')))
+        ))
         self.assertEqual(pkg.link_options(self.linker), option_list(
-            '/LIBPATH:/path', opts.lib_literal('foo.lib'),
+            '/DEBUG', opts.lib_dir(Directory(Path('/path'))),
+            opts.lib_literal('foo.lib'),
             (opts.rpath_dir(Path('/path')) if self.platform_name == 'linux'
              else None)
         ))